+++ packages/wiz-ui-react/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @wizleap-inc/wiz-ui-react
 
+## 0.57.0
+
+### Minor Changes
+
+- [#1293](https://github.com/Wizleap-Inc/wiz-ui/pull/1293) [`f2a295ce`](https://github.com/Wizleap-Inc/wiz-ui/commit/f2a295cee04e680b1b2cfb35be955eafea65edf1) Thanks [@RyushiAok](https://github.com/RyushiAok)! - Feat(icon): WizIDelete, WizIReviews
+
+### Patch Changes
+
+- Updated dependencies [[`f2a295ce`](https://github.com/Wizleap-Inc/wiz-ui/commit/f2a295cee04e680b1b2cfb35be955eafea65edf1)]:
+  - @wizleap-inc/wiz-ui-constants@0.24.0
+  - @wizleap-inc/wiz-ui-styles@0.29.5
+
+## 0.56.0
+
+### Minor Changes
+
+- [#1284](https://github.com/Wizleap-Inc/wiz-ui/pull/1284) [`d8e63723`](https://github.com/Wizleap-Inc/wiz-ui/commit/d8e6372313d7e1ccba97c266166d829a148e112a) Thanks [@RyushiAok](https://github.com/RyushiAok)! - Feat(icons): emergency_home, hourglass_top, feedback, menu_open アイコンを追加
+
+### Patch Changes
+
+- [#1277](https://github.com/Wizleap-Inc/wiz-ui/pull/1277) [`7c97d8b7`](https://github.com/Wizleap-Inc/wiz-ui/commit/7c97d8b791ccc73df783f68620d1ee18b9d6cdc7) Thanks [@RyushiAok](https://github.com/RyushiAok)! - Fix(text-button): 横幅によって Icon の下に Text が回り込む問題を修正
+
+- Updated dependencies [[`d8e63723`](https://github.com/Wizleap-Inc/wiz-ui/commit/d8e6372313d7e1ccba97c266166d829a148e112a)]:
+  - @wizleap-inc/wiz-ui-constants@0.23.0
+  - @wizleap-inc/wiz-ui-styles@0.29.4
+
+## 0.55.0
+
+### Minor Changes
+
+- [#1275](https://github.com/Wizleap-Inc/wiz-ui/pull/1275) [`2fc87c27`](https://github.com/Wizleap-Inc/wiz-ui/commit/2fc87c277cb125ff6d6860018f073d8d0ca46c5f) Thanks [@RyushiAok](https://github.com/RyushiAok)! - Feat(nav-item): 外部リンクを新しいタブで開く
+
+## 0.54.0
+
+### Minor Changes
+
+- [#1266](https://github.com/Wizleap-Inc/wiz-ui/pull/1266) [`75eeecde`](https://github.com/Wizleap-Inc/wiz-ui/commit/75eeecdec6559df41af8f1080121db4da53f4377) Thanks [@RyushiAok](https://github.com/RyushiAok)! - Feat(chat-card): 開閉状態を指定可能にする
+
+## 0.53.2
+
+### Patch Changes
+
+- [#1262](https://github.com/Wizleap-Inc/wiz-ui/pull/1262) [`06f21fe0`](https://github.com/Wizleap-Inc/wiz-ui/commit/06f21fe0f979f3f6c7a86e51b45843d1001f4ff2) Thanks [@RyushiAok](https://github.com/RyushiAok)! - Fix(PollGraph): 0%での表示を Vue 版と揃える
+
+- Updated dependencies [[`80a88bf3`](https://github.com/Wizleap-Inc/wiz-ui/commit/80a88bf39b4d378b4357e2531525561e574940b3)]:
+  - @wizleap-inc/wiz-ui-styles@0.29.3
+
 ## 0.53.1
 
 ### Patch Changes